refactor: Set horizontal cells to 1 on both axes
LODs now generate chunks with 1 cell on both horizontal axes (X and Z), and some more on the vertical axis (Y). This means the chunks will generate with 1 sample on the horizontal axes, massively improving performance at the cost of some LOD fidelity.
Also improved an if statement by a negligible amount by removing the `=` check.